Add format_json template filter
